+++ lib/trailblazer/operation/nested.rb
@@ -6,19 +6,22 @@ class Trailblazer::Operation
   end
 
   module Nested
-    def self.proc_for(step)
+    # _call the nested `operation`.
+    Call = ->(operation, options) { operation._call(*options.to_runtime_data) }
+
+    def self.strut_for(step)
       if step.is_a?(Class) && step <= Trailblazer::Operation # interestingly, with < we get a weird nil exception. bug in Ruby?
-        proc = ->(input, options) { step._call(*options.to_runtime_data) }
+        strut = ->(input, options) { Call.(step, options) }
       else
         option = Option::KW.(step)
-        proc = ->(input, options) { option.(input, options).(*options.to_runtime_data) }
+        strut = ->(input, options) { Call.(option.(input, options), options) }
       end
     end
 
     # Please note that the instance_variable_get are here on purpose since the
     # superinternal API is not entirely decided, yet.
     def self.for(step)
-      proc = proc_for(step)
+      proc = strut_for(step)
 
       ->(input, options) do
         result = proc.(input, options) # TODO: what about containers?