+++ CHANGES.md
@@ -1,6 +1,7 @@
 # 2.0.3
 
 * `Guard` now allows kw args for its option.
+* Fix a bug where `Nested( ->{} )` wouldn't `_call` the nested operation and did too much work on re-nested the already nested params. Thanks to @eliranf for spotting this.
 
 # 2.0.2